+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 0.1.0 – *03/14/2017*
+# 0.2.0
+
+> *16/07/2017*
+
+- Caches now support arbitrary keys and namespacing. This is done via the `CacheKey` trait, that
+  represent a bijection between keys and values.
+
+# 0.1.0
+
+> *14/03/2017*
+
 
 - Initial revision.